+++ packages/agent-core/test/tools/cron/cron-delete.test.ts
@@ -1,35 +1,11 @@
 /**
- * Tests for `tools/cron/cron-delete.ts`.
- *
- * Same harness as `cron-create.test.ts` / `cron-list.test.ts`: a tiny
- * Agent stub exposes only `turn` + `telemetry` (the slices the manager
- * actually touches), and `ClockSources` is injected so nothing in the
- * test relies on the system clock.
- *
- * The behaviours under test mirror the contract documented in
- * `cron-delete.ts`:
- *
- *   1. A real deletion drains the store, returns a success message, and
- *      emits exactly one `cron_deleted` telemetry event keyed on the
- *      task id.
- *   2. A delete of a well-formed-but-missing id is reported as an error
- *      (isError: true) with a "no cron job with id …" message, and
- *      emits no telemetry.
- *   3. Malformed ids (uppercase, too short, non-hex, empty) never reach
- *      the store; the tool returns an error that names the constraint.
- *      No store mutation, no telemetry.
- *
- * These cases together pin the "report-and-correct" contract: the model
- * sees a failure on every code path that would otherwise be a silent
- * no-op.
+ * Tests for `tools/cron/cron-delete.ts`. Pins the report-and-correct
+ * contract: every code path that would otherwise be a silent no-op
+ * (missing id, malformed id) reports an error.
  */
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import type { ContentPart } from '@moonshot-ai/kosong';
 
-import type { Agent } from '../../../src/agent';
-import type { PromptOrigin } from '../../../src/agent/context/types';
 import { CronManager } from '../../../src/agent/cron/manager';
-import type { ClockSources } from '../../../src/tools/cron/clock';
 import {
   CronDeleteTool,
   type CronDeleteInput,
@@ -41,58 +17,11 @@ import type {
   RunnableToolExecution,
   ToolExecution,
 } from '../../../src/loop/types';
-
-// Same wall anchor as the sibling cron tests for cross-file
-// comparability of any timing-derived assertions.
-const WALL_ANCHOR = 1_700_000_000_000;
-
-interface SteerCall {
-  readonly content: readonly ContentPart[];
-  readonly origin: PromptOrigin;
-}
-interface TelemetryCall {
-  readonly event: string;
-  readonly props: unknown;
-}
-
-interface AgentStub {
-  readonly agent: Agent;
-  readonly steerCalls: SteerCall[];
-  readonly telemetryCalls: TelemetryCall[];
-}
-
-function createAgentStub(): AgentStub {
-  const steerCalls: SteerCall[] = [];
-  const telemetryCalls: TelemetryCall[] = [];
-  const turn = {
-    get hasActiveTurn(): boolean {
-      return false;
-    },
-    steer: (content: readonly ContentPart[], origin: PromptOrigin) => {
-      steerCalls.push({ content, origin });
-      return 42;
-    },
-  };
-  const telemetry = {
-    track: (event: string, props: unknown) => {
-      telemetryCalls.push({ event, props });
-    },
-  };
-  const agent = { turn, telemetry } as unknown as Agent;
-  return { agent, steerCalls, telemetryCalls };
-}
-
-/**
- * Frozen-wall clock. CronDelete does not read `wallNow()` itself, but
- * the manager's `isStale` path (unused here) and `store.add(nowMs, …)`
- * (used during seeding) do — so we pin one value and reuse it.
- */
-function createClocks(wall = WALL_ANCHOR): ClockSources {
-  return {
-    wallNow: () => wall,
-    monoNowMs: () => 1_000_000,
-  };
-}
+import {
+  createAgentStub,
+  createClocks,
+  type AgentStub,
+} from '../../agent/cron/harness/stub';
 
 interface Harness {
   readonly stub: AgentStub;
@@ -103,7 +32,7 @@ interface Harness {
 function makeHarness(): Harness {
   const stub = createAgentStub();
   const manager = new CronManager(stub.agent, {
-    clocks: createClocks(),
+    clocks: createClocks().clocks,
     pollIntervalMs: null,
   });
   const tool = new CronDeleteTool(manager);
